feat: Add First-In First-Out (FIFO) caching module
The code changes introduce a new file `0x01-caching/1-fifo_cache.py` which implements a First-In First-Out (FIFO) caching module. This module allows storing and retrieving items from a dictionary with a FIFO removal mechanism when the cache limit is reached. The `put` method adds an item to the cache, and if the cache limit is exceeded, the oldest item is discarded. The `get` method retrieves an item by key from the cache. This FIFO caching functionality will improve the performance and efficiency of the application.