feat: Add LIFO (Last-In First-Out) caching module
The code changes introduce a new file `0x01-caching/2-lifo_cache.py` which implements a Last-In First-Out (LIFO) caching module. This module allows storing and retrieving items from a dictionary with a LIFO removal mechanism when the cache limit is reached. The `put` method adds an item to the cache, and if the cache limit is exceeded, the oldest item is discarded. The `get` method retrieves an item by key from the cache. This LIFO caching functionality will improve the performance and efficiency of the application.