feat: Implement LRUCache class for Least Recently Used caching
The code changes add a new file `0x01-caching/3-lru_cache.py` which implements the LRUCache class for Least Recently Used caching. This class allows storing and retrieving items from a dictionary with a LRU removal mechanism when the cache limit is reached. The `put` method adds an item to the cache, and if the cache limit is exceeded, the least recently used item is discarded. The `get` method retrieves an item by key from the cache. This LRUCache functionality will improve the performance and efficiency of the application.