feat: Add MRUCache class for Most Recently Used caching
The code changes add a new file `0x01-caching/4-mru_cache.py` which implements the MRUCache class for Most Recently Used caching. This class allows storing and retrieving items from a dictionary with an MRU removal mechanism when the cache limit is reached. The `put` method adds an item to the cache, and if the cache limit is exceeded, the most recently used item is discarded. The `get` method retrieves an item by key from the cache. This MRUCache functionality will improve the performance and efficiency of the application.