feat: Add internationalization support to Flask app
The code changes add internationalization support to the Flask app by introducing a new configuration file `babel.cfg` and translation files `messages.pot`, `messages.po` for English and French languages. This allows for the localization of the app's content based on the user's preferred language. The addition of these files lays the foundation for developing a multilingual Flask application.